+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.0.0-alpha.3] - 2018-08-05
+### Added
+- `@material-ui` packages to give a Material Design feel to the component.
+
+### Changed
+- Allow selection of multiple items
+
 ## [1.0.0-alpha.2] - 2018-05-01
 ### Added
 - `.npmignore` file to reduce package size.